requirements.txt
appdirs>=1.4.4
setuptools>=77.0.3
cython>=3.0.0
wheel>=0.41.0
ninja>=1.11.1
pyinstaller>=5.13.0
pysdl2>=0.9.14
pysdl2-dll>=2.26.0
vulkan>=1.3.236
